backend/requirements.txt
# Core dependencies
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
pillow>=10.2.0
numpy>=1.26.2

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.1.2

# AI-based classification (CLIP)
transformers>=4.35.2
torch>=2.1.1

# Email validation
pydantic[email]
